Changed ":quot" to ":quote".
https://github.com/umr4nlp/umr-guidelines/issues/22
jbonn : "Yes, there were a number of last minute changes that were made in post processing for the 1.0 release in the name of making the graphs more human-understandable. I didn't notice that this was one of them so thank you for pointing it out! All of these need to be changed in the guidelines, editor, and various lists now. (Others in this same vein included changing :modstr to :modstrength-- I think-- and changing the values to full-affirmative instead of FullAff, etc.)"